refactor(ui): split APITokenManager into list/create/usage sub-components
